### 2019-11-22 #72 @whedon
[PRE REVIEW]: EduCortex: browser-based 3D brain visualization of fMRI meta-analysis maps
**Submitting author:** @PaulScotti (<a href="http://orcid.org/0000-0003-4912-8809">Paul Scotti</a>)
**Repository:** <a href="https://github.com/PaulScotti/EduCortex" target ="_blank">https://github.com/PaulScotti/EduCortex</a>
**Version:** v1.0.0
