### 2019-12-02 #72 @PaulScotti
Could someone let me know why whedon is unable to find any papers to compile? Is it because the repository listed is [github.com/PaulScotti/EduCortex](url) and not [](https://github.com/PaulScotti/EduCortex) ?
