### 2019-12-02 #72 @kyleniemeyer
I think the repo URL is configured with JOSE/whedon incorrectly, since the link above directs to https://github.com/openjournals/jose-reviews/issues/github.com/PaulScotti/EduCortex (rather than https://github.com/PaulScotti/EduCortex)
