### 2019-12-02 #72 @kyleniemeyer
@openjournals/dev is it possible to change the repo URL? I don't see an option in the editors dashboard
