+++ paper/paper.md
@@ -49,6 +49,9 @@ Below we detail the features of EduCortex in line with the steps displayed in th
 
 ![EduCortex Diagram](figure1.png) Figure 1. Image depicting what EduCortex initially looks like upon visiting the website. Dotted boxes and numbers are not displayed in the tool and are depicted here purely for the illustrative purpose of outlining EduCortex's features.
 
+# Statement of Need
+EduCortex is an educational tool that can be accessed through any web browser with an internet connection. We believe that this tool will be useful for instructional purposes. As an example, if a student wonders "what part of the brain processes words?", then they can search for "word" and see activation predominantly in the visual word form area located on the left fusiform gyrus [@mccandliss2003], in addition to more distributed areas of the brain. The reverse process of clicking a brain region to view the most associated terms allows the user to freely explore the brain to understand functional specialization. An improved anatomical understanding of the brain can also be gleaned by rotating, inflating, and flattening the brain in addition to the anatomical label overlay provided in the viewer. The principal components analysis maps illustrate the large-scale functional specialization of the brain by depicting the terms that explained the most variance across NeuroSynth activation maps. 
+
 # Features
 ## 1. Interactive 3D Brain
 The brain used for all visualization is the "fsaverage" brain taken from Freesurfer [@freesurfer]. The brain can be rotated by clicking and dragging with the mouse. It can also be resized by right-clicking and dragging with the mouse. The initial activation map displayed on the brain is a colormap representing the top 3 principal components of our dimensionality reduction procedure (see Principal Components Analysis section). 
