+++ paper/paper.bib
@@ -37,3 +37,13 @@
   year={2016},
   publisher={Nature Publishing Group}
 }
+@article{mccandliss2003,
+  title={The visual word form area: expertise for reading in the fusiform gyrus},
+  author={McCandliss, Bruce D and Cohen, Laurent and Dehaene, Stanislas},
+  journal={Trends in cognitive sciences},
+  volume={7},
+  number={7},
+  pages={293--299},
+  year={2003},
+  publisher={Elsevier}
+}
