### 2019-12-04 #72 @arfon
@PaulScotti - you can ask Whedon to generate a new PDF preview by typing:
