### 2020-01-23 #72 @magsol
Yes, @PaulScotti, very sorry about the delay. I'm still getting my semester up and running but given a reviewer or two I have in mind who specialize in computational neuroscience, plus the very helpful list you provided me, I'm hoping we can have things going in the next couple days.
