### 2020-03-04 #75 @tliu23
Went through the lists and checked the ones.
