### 2020-03-05 PaulScotti/educortex#5 @TomDonoghue
Interface Bug: 'Data' selector in the top left interaction box
When I interact with the viewer on the website, I see an interaction panel in the top left, that lets the user click between the search term results and the clickable viewer. It also has an option to click on data, however, whenever I click on that I get a strippy pattern across the brain, and the brain viewer is usually also cut off.

For example:
![Screen Shot 2020-03-05 at 2 45 31 PM](https://user-images.githubusercontent.com/7727566/76033013-ff33c900-5eef-11ea-8058-97746dc7d6f3.png)

I'm not sure I understand what the data tab is actually supposed to do. Either this could be clarified somewhere, and fixed, or perhaps this option could just be removed?

As a more minor point, I am unsure why this selector has up and down arrow indicators, suggesting there is a drop-down list, but there does not seem to be any drop downs, and it doesn't seem like the arrows are needed.

Another minor point, but I'm not sure the labels on the other two options are super clear, and they could perhaps be renamed (maybe to something like "Search View" and "Click View" - though I imagine better options could be thought of).

I checked that this issue happens across Firefore, Chrome, and Safari.

Note: this issue filed as part of doing a review for JOSE (https://github.com/openjournals/jose-reviews/issues/75)
