### 2020-03-06 PaulScotti/educortex#5 @PaulScotti
Thanks for catching this, that panel wasn't actually supposed to be clickable. I updated it so that the panel isn't clickable.
