### 2020-03-07 PaulScotti/educortex#6 @TomDonoghue
Install Instructions
For the install instructions, they are (I think) currently a copy of some old instructions for installing `pycortex` (also, there is a typo in the line `python setup.py developpip install`). The instructions say to then proceed to use `runMappingViz.ipynb`, but this notebook has a series of extra dependencies over and above `pycortex` (including, for example `neurosynth` and `sklearn` - so without saying anything about these, the instructions are not sufficient to be able to run this project locally.

My recommendation would be to:
- add a `requirements.txt` file, listing all dependencies, including `pycortex`, and what is needed to run the notebook. Install instructions can then say to install the dependencies listed in `requirements.txt`
- for pycortex itself (a more complex install), either just direct to the install instructions on their page, or at least update to the current instructions in their README, which allow for installing a stable version through from pypi.

There could also be other strategies to updating this - but I don't think the current instructions are ideal.

Note: this issue filed as part of doing a review for JOSE (openjournals/jose-reviews#75)
