### 2020-03-07 PaulScotti/educortex#7 @TomDonoghue
Suggestion: Add Versioning
There is currently no versioning, or tagged release structure for this project. Since this project does indicate it could accept contributions with changes, and updates, it seems like it would be useful to tag a version number to the project, and make release tags with any bigger updates.

I know this isn't exactly a module, but my best guess is that a `version.py` file would suffice for this.

Note: this issue filed as part of doing a review for JOSE (openjournals/jose-reviews#75)
(This came to mind partly as I'm pretty sure you will need a version number to tag a release with publication, and so it might as well have a more standard version tagging.
