### 2020-03-08 PaulScotti/educortex#8 @TomDonoghue
Suggestion: General Audience Description
My sense is that the goal of this project is as an educational tool for users to explore  (and I really like this idea). However, it seems like the intended audience may not necessarily already know details about brain imaging and in particular, neurosynth (and I presume the idea is not that they must read the more technical overview in JOSE paper or understand the neurosynth tool to be able to get a sense of what they are looking at). It seems to me that, pedagogically, this tool would be all the most useful if it leverages the cool interactive tool in a way that is grounded in the realities of the science it reflects, and offers straight-forward information on how to use it and think about it.

My thought then is to suggest adding some high level description for the potential viewer of how this works, and what it means. This could be "How it works" link from the site with, for example, just a paragraph or two that briefly and accessibly describes the data that goes into this, what it means, and how it can be interpreted.

I think this might be all the more important with the PCA results. I like the idea of showing what topics explain the most variance, but without a bit more context of description, I think it's likely a bit difficult for the casual user to understand and fruitfully interact with the bottom right part of the page. It took me a while to understand the PCA visualizations, as the colormap on the brain, and the plotted wordcloud, and when and why this disappeared. I don't think the wordcloud, or why different terms are coloured in in the list at the bottom as you click through the different PCA labels is likely to be very clear to the potential audience, and guide to what this is, and how to interact with it may be especially useful.

Note: this issue filed as part of doing a review for JOSE (openjournals/jose-reviews#75)
