### 2020-03-08 PaulScotti/educortex#11 @TomDonoghue
Suggestion: Toggle PCA Wordcloud Viz On/Off
Could there be a toggle to turn the PCA wordcloud on & off, just like how you can toggle the controls and associated papers displays on or off?

The reason is, when I turn on the ROI mapping on, and zoom into the brain to explore, the PCA wordcloud is in the way and overlaps on top of the brain in a sort of annoying manner.
