### 2020-03-08 #75 @TomDonoghue
Okay, sorry for the delay, I have worked through the project and had some time to work on my review. I opened issues on the original repo with more details, with the main points recapped here.

Main Review Comments:
- [ ] I think the dependencies and install instructions can be updated  (https://github.com/PaulScotti/educortex/issues/6)
- [ ] I think it might be useful to tag a versions of the site (https://github.com/PaulScotti/educortex/issues/7)
- [ ] I recommend the site include or link to some brief descriptions on how this thing works, so it can be appropriately interpreted as an educational tool(https://github.com/PaulScotti/educortex/issues/8)
- [ ] I think some organization and description of the repository would be useful (https://github.com/PaulScotti/educortex/issues/9)
    - I added notes here about possible re-organizations of the module, which is broadly up to the authors how much they want to do this. My main concern is making sure the tool is structured to ensure it can be maintained, others could potentially contribute, and be stable so that it can stay online.
- [ ] Whedon has indicated above that there are 2 missing DOIs that need adding to the references in the paper.

Overall, and to be clear, I do think this tool / website is really cool! The issues I opened also include broader suggestions than I would consider necessary for me signing off on the review - I just wanted to lay out my suggestions.
