### 2020-03-11 f1083b2
added requirements and changed install instructions
