+++ README.md
@@ -5,11 +5,11 @@
 
 ## Overview
 
-[EduCortex](https://paulscotti.com/educortex/) is an educational browser-based visualization tool that allows the user to enter any functional or anatomical term (e.g., "visual", "face", "motion", "precuneus") and visualize the parts of the brain that are most associated with that term. The process can also be reversed, where the user can click anywhere on the brain to see what terms are most associated with the selected brain region. Using principal components analysis, we also display the functional terms that explained the most variance across all activation maps. This tool works through the combination of [Neurosynth](http://neurosynth.org/), a large-scale, automated database of fMRI papers, and [PyCortex](https://gallantlab.github.io/), an interactive 3D brain visualizer.
+[EduCortex](http://paulscotti.com/educortex/) is an educational browser-based visualization tool that allows the user to enter any functional or anatomical term (e.g., "visual", "face", "motion", "precuneus") and visualize the parts of the brain that are most associated with that term. The process can also be reversed, where the user can click anywhere on the brain to see what terms are most associated with the selected brain region. Using principal components analysis, we also display the functional terms that explained the most variance across all activation maps. This tool works through the combination of [Neurosynth](http://neurosynth.org/), a large-scale, automated database of fMRI papers, and [PyCortex](https://gallantlab.github.io/), an interactive 3D brain visualizer.
 
 ## Installation
 
-EduCortex can be accessed without installation as an online webviewer, see [EduCortexViewer](https://paulscotti.com/educortex/).
+EduCortex can be accessed without installation as an online webviewer, see [EduCortexViewer](http://paulscotti.com/educortex/).
 
 If you want to use EduCortex locally, take the following steps:
 
