### 2020-03-12 ca4b136
add versioning, more detailed instructions
