+++ README.md
@@ -1,4 +1,4 @@
-# EduCortex (v.1.0.0)
+# EduCortex
 <img src="logo.png" align="center"/>
 
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
