### 2020-03-12 cadb0c1
add doi to references of jose paper
