+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={8},
   pages={665},
   year={2011},
+  doi={10.1038/nmeth.1635},
   publisher={Nature Publishing Group}
 }
 @article{pycortex,
@@ -15,6 +16,7 @@
   volume={9},
   pages={23},
   year={2015},
+  doi={10.3389/fninf.2015.00023},
   publisher={Frontiers}
 }
 @article{freesurfer,
@@ -25,6 +27,7 @@
   number={2},
   pages={774--781},
   year={2012},
+  doi={10.1016/j.neuroimage.2012.01.021},
   publisher={Elsevier}
 }
 @article{glasser2016,
@@ -35,6 +38,7 @@
   number={7615},
   pages={171},
   year={2016},
+  doi={10.1038/nature18933},
   publisher={Nature Publishing Group}
 }
 @article{mccandliss2003,
@@ -45,5 +49,6 @@
   number={7},
   pages={293--299},
   year={2003},
+  doi={10.1016/S1364-6613(03)00134-7},
   publisher={Elsevier}
 }
