### 2020-03-12 PaulScotti/educortex#11 @PaulScotti
Added a button to the bottom right that allows turning the wordcloud on and off
