### 2020-03-12 PaulScotti/educortex#8 @PaulScotti
Added a simple welcome message to the tool as well as a "what is this" button at the bottom right to describe the wordcloud/PCA. We also have an idea in the works for a future update to better convey the high-level ideas behind the tool.
