### 2020-03-12 PaulScotti/educortex#7 @PaulScotti
Added versioning via git release: https://github.com/PaulScotti/educortex/releases
