### 2020-03-12 PaulScotti/educortex#6 @PaulScotti
Added requirements.txt, changed our readme file to better convey install instructions, and edited the runMappingViz file to improve readability.
