### 2020-03-12 PaulScotti/educortex#9 @PaulScotti
I appreciate the suggestion and hope that we can improve this repository in the ways you suggest in the future. For the moment it would take considerable time to make these changes, although I did go ahead and revise the runMappingViz file to improve readability in terms of how we did PCA and create the functional png maps extracted from Neurosynth. Our primary goal for now is to disseminate this tool (aka the educortex website) to teachers and children rather than work towards improving the ease with which contributors can help improve the tool.
