### 2020-03-12 #75 @PaulScotti
Thanks everyone for their time and effort spent reviewing this project. Thanks especially to @TomDonoghue for the detailed suggestions. I have made revisions to address the issues/suggestions raised, including

- fixed two bugs: the data selector (top-left panel) is now not clickable, and the pca visualization now correctly switches for "Top 7-9"

- Added versioning using git tag (starting now at v1.0)

- Added welcome message that is presented when the webpage is first loaded

- Added "what is this?" button next to the wordcloud that describes the wordcloud/PCA in a hopefully easy to digest manner

- Added a "disable wordcloud" button next to the "what is this?" button that allows the user to turn off the wordcloud while the PCA visualization is still active

- Revised installation instructions in the README to correctly inform the reader about dependencies

- Revised runMappingViz.ipyb to detail how we created the Neurosynth term maps and PCA maps with better commenting to help the reader render their own local server of EduCortex

- Revised manuscript to include dois for references
