### 2020-03-14 PaulScotti/educortex#11 @TomDonoghue
Hey @PaulScotti - thanks for implementing this, just a note that I don't seem to see a way to turn the wordcloud back on..? When I click the wordcloud off, the button disappears, so it can't be clicked again. Can the button be set to stay there, just like the "Associated Studies" and "Open Controls" ones do?
