### 2020-03-14 PaulScotti/educortex#6 @TomDonoghue
Hey @PaulScotti, I think the requiremnts file is good, but does it need to be so specific, in terms of exact version requirements? This looks like it comes from an environment description, but presumably any new versions of at least most of these requirements are fine, and it means it wouldn't make local setups have to change to these specific versions if they have newer version.

Also, the readme could also note that you need to install requirements, so maybe step 2 could note something like:
```
# move into the cloned repository
cd educortex
# install the list of required dependencies
pip install -r requirements.txt
```
