### 2020-03-14 PaulScotti/educortex#9 @TomDonoghue
I get that bigger refactors could be a fair amount of work - I think they don't need to happen now, but maybe could be kept in mind as a potential development milestone.

I also like that you updated `runMappingViz` but I still feel it's missing an intro - could you perhaps add a markdown cell at the top with a couple sentences on what it does?

As a quick addition, what I think would help the incomer navigate better, what about just a brief 'repository layout' description in the README? This, I think, would lay out the top level logic in a helpful manner. I'm thinking of something like this:
https://github.com/voytekresearch/BandRatios/blob/master/README.md#repository-layout
