### 2020-03-14 #75 @TomDonoghue
Thanks for all the updates @PaulScotti

Just a note that I did leave follow up comments on a couple issues on educortex (https://github.com/PaulScotti/educortex/issues/6 & https://github.com/PaulScotti/educortex/issues/11) that you've been working on.

For the purpose of the review, I would say I'm pretty much happy with it all. In terms of broader organizations, I added a comment to the relevant issue (https://github.com/PaulScotti/educortex/issues/9) that I think a brief 'Repository Layout' would help people interact with the repo, and potentially contribute.

As I double check the review points, also if the idea is that people should use the Github issue tracker for filing bugs / asking questions (as opposed to other options), perhaps that should be listed in the README?

I didn't have any big comments about the paper itself - it's seem a nice and straight-forward description of how this tool works. My only thought was if maybe the paper should say a bit more about motivating the statement of need (why is it useful to learn about functional neuroanatomy?), but I'm less sure about how much the paper should address this, so maybe @magsol has thoughts on that.

Otherwise, after these last couple things, all looking good to me!
