### 2020-03-18 PaulScotti/educortex#11 @PaulScotti
Ah thanks, yes I added an enable button. The PCA/wordcloud can also be reenabled by clicking the "Top x-y" button under the camera viewer.
