+++ requirements.txt
@@ -1,12 +1,12 @@
-cloudpickle==1.1.1
-matplotlib==3.1.1
-neurosynth==0.3.8
-nibabel==2.5.0
-nilearn==0.6.2
-numpy==1.16.5
-pandas==0.25.2
-pycortex==1.1.1
-sklearn==0.0
-scikit-learn==0.21.3
-scikit-optimize==0.5.2
-scipy==1.2.1
+matplotlib
+neurosynth
+nibabel
+nilearn
+numpy
+pandas
+pycortex
+cloudpickle
+sklearn
+scikit-learn
+scikit-optimize
+scipy
