### 2020-03-18 9f8ceab
better descriptions for README and first cell in runMappingViz, and remove redundant wordclouds folder
