### 2020-03-18 05fd243
minor formatting changes to README
