### 2020-03-18 PaulScotti/educortex#6 @PaulScotti
Thanks, made the suggested changes. Sorry, the reason I was so specific is that I'm not very familiar with typical requirements.txt files, I thought it was supposed to be the exact version numbers in case later updates to a package lead to errors in the code. Thanks for the suggestions.
