### 2020-03-18 PaulScotti/educortex#9 @PaulScotti
Added an intro to runMappingViz and also repository layout to the README (thanks for the reference!). And yes, your above suggestions are definitely something I wish we can do as a potential developmental milestone.
