### 2020-03-18 7efeba3
added statement about filing bugs and asking questions using github issue tracker
