### 2020-03-18 PaulScotti/educortex#6 @TomDonoghue
I think of a `requirements.txt` as a minimal requirements - so just module names, and any minimum versions, if there are any. But this is then more general, easier, than an environment file, which is what I think you had before basically, which a specific environment, with everything specified, which is sorta like an 'existence proof' of an environment that works. For install, I think one generally wants the general version, unless there is some reason one wants to mandate an exact overall environment (This is just my sense - not necessarily strict guidelines).

Anyways - thanks for updating!
