### 2020-03-18 #75 @TomDonoghue
With the last round of updates that @PaulScotti put through, I would say I have covered all the things that I had for my review. Unless their are any further thoughts / direction about the paper itself, I'd say I'm good to sign off on this review.

Overall: I'd say this is a cool tool, and I really like the website. I think it makes for a good JOSE paper. Congratulations to the developers for this work, and I hope you continue to work on and maintain this tool (and maybe modularize as you go) to keep this website up and running.

@magsol Let me know if you need anything else from me! Thanks all!
