### 2020-03-28 #75 @magsol
Sorry for the delay--university changes around COVID-19 have upended my schedule from its roots, though I'm sure I don't have to explain that to anyone here.

Thanks @tliu23 and @TomDonoghue for your reviews, and especially Tom for the insightful comments and feedback. This was a lot of fun to read, and super awesome that this came out of Neurohackademy; it's such a phenomenal annual workshop/meeting.

To finalize EduCortex for publication, @PaulScotti I need you to make a tagged release & archive of the artifact. Once that's done, report version number and DOI in this thread, and we can take it from there.
