### 2020-03-29 89ce6ec
add version and jose buttons to readme
