### 2020-03-29 #75 @PaulScotti
v1.1: https://github.com/PaulScotti/educortex/releases/tag/v1.1
archive: https://doi.org/10.5281/zenodo.3732375

Huge thanks to @magsol @TomDonoghue and @tliu23 for their time and effort spent facilitating this review process, I'm very thankful that this journal exists to support projects like this.
