### 2020-04-03 #75 @magsol
Hi @openjournals/jose-eics , whenever you get a chance, please give @PaulScotti 's submission here a final yay/nay!
