### 2020-04-04 #75 @labarba
@PaulScotti — Could you edit the metadata of the Zenodo deposit so the title and author list matches the paper? Thanks!
