### 2020-04-04 #75 @PaulScotti
@labarba I corrected the author names/affiliations and the title of the Zenodo deposit, are there other metadata that I should correct?
